feat: agregar reinicio de contraseña y ver perfil

+++ src/main/java/com/perfulandia/usuarios/repository/UsuarioRepository.java
@@ -1,11 +1,9 @@
 package com.perfulandia.usuarios.repository;
 
+import com.perfulandia.usuarios.model.entity.Usuario;
+import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.perfulandia.usuarios.model.entity.Usuario;
-
-public interface UsuarioRepository extends JpaRepository<Usuario, Integer> {
+public interface UsuarioRepository extends JpaRepository<Usuario, Long> {
     Optional<Usuario> findByCorreo(String correo);
 }
